cards

+++ controllers/cards.js
@@ -5,21 +5,24 @@ const Card = require('../models/card');
 const { BadRequestError } = require('../errors/BadRequestError');
 const { InternalServerError } = require('../errors/InternalServerError');
 const { NotFoundError } = require('../errors/NotFoundError');
+const { ForbiddenError } = require('../errors/Forbidden');
 
-// Получить все карты
+// Получить все карточки
 module.exports.getCards = async (req, res, next) => {
   try {
     const cards = await Card.find();
+    if (!cards) {
+      throw new NotFoundError('Карточки не найдены');
+    }
     res.status(200).send({ data: cards });
   } catch (error) {
     next(error);
   }
 };
 
-// Создать новую карту
 module.exports.createCard = async (req, res, next) => {
   try {
-    console.log(req.user._id); // _id будет доступен
+    console.log(req.user._id);
     const { name, link } = req.body;
 
     if (!name || !link) {
@@ -27,30 +30,37 @@ module.exports.createCard = async (req, res, next) => {
     }
 
     const card = await Card.create({ name, link, owner: req.user._id });
-    res.status(201).send({ data: card });
+    res.status(201).send({ data: card._id });
   } catch (error) {
     if (error.name === 'ValidationError') {
-      next(new BadRequestError(error.message));
+      next(new BadRequestError('Неверный ID карточки'));
     } else {
       next(new InternalServerError('Произошла ошибка на сервере'));
     }
   }
 };
 
-// Удалить карту
+// Удалить карточку
 module.exports.deleteCard = async (req, res, next) => {
   try {
     const { cardId } = req.params;
+    const { userId } = req.user; // Идентификатор текущего пользователя
 
     if (!mongoose.Types.ObjectId.isValid(cardId)) {
-      throw new BadRequestError('Неверный ID картинки');
+      throw new BadRequestError('Неверный ID карточки');
     }
 
-    const card = await Card.findByIdAndRemove(cardId);
+    const card = await Card.findById(cardId);
 
     if (!card) {
-      throw new NotFoundError('Картинка не найдена');
+      throw new NotFoundError('Карточка не найдена');
     }
+
+    if (card.userId !== userId) { // Проверка, что карточка принадлежит текущему пользователю
+      throw new ForbiddenError('У вас нет доступа к этой карточке');
+    }
+
+    await Card.findByIdAndRemove(cardId);
 
     res.status(200).send({ data: card });
   } catch (error) {
@@ -64,7 +74,7 @@ module.exports.likeCard = async (req, res, next) => {
     const { cardId } = req.params;
 
     if (!mongoose.Types.ObjectId.isValid(cardId)) {
-      throw new BadRequestError('Неверный ID картинки');
+      throw new BadRequestError('Неверный ID карточки');
     }
 
     const card = await Card.findByIdAndUpdate(
@@ -74,10 +84,10 @@ module.exports.likeCard = async (req, res, next) => {
     );
 
     if (!card) {
-      throw new NotFoundError('Картинка не найдена');
+      throw new NotFoundError('Карточка не найдена');
     }
 
-    res.status(200).send({ data: card });
+    res.status(200).send({ data: card.likes });
   } catch (error) {
     next(error);
   }
@@ -89,7 +99,7 @@ module.exports.deleteLike = async (req, res, next) => {
     const { cardId } = req.params;
 
     if (!mongoose.Types.ObjectId.isValid(cardId)) {
-      throw new BadRequestError('Неверный ID картинки');
+      throw new BadRequestError('Неверный ID карточки');
     }
 
     const card = await Card.findByIdAndUpdate(
@@ -99,10 +109,10 @@ module.exports.deleteLike = async (req, res, next) => {
     );
 
     if (!card) {
-      throw new NotFoundError('Картинка не найдена');
+      throw new NotFoundError('Карточка не найдена');
     }
 
-    res.status(200).send({ data: card });
+    res.status(200).send({ data: card.likes }); // исправление: возвращаем только список лайков
   } catch (error) {
     next(error);
   }
